fix: rename apply-supplier to ApplySupplier
